# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TLITODOS Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/users/me/notifications": {
      "patch": {
        "tags": [
          "authorization"
        ],
        "summary": "Update Notifications",
        "operationId": "update_notifications_api_v1_users_me_notifications_patch",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationSettingsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/notifications": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "List Notifications",
        "operationId": "list_notifications_api_v1_notifications_get",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "TODO_COMPLETED",
                    "DIARY_CREATED",
                    "BET_REQUESTED"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationsPageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/{notificationId}/read": {
      "patch": {
        "tags": [
          "notifications"
        ],
        "summary": "Read Notification",
        "operationId": "read_notification_api_v1_notifications__notificationId__read_patch",
        "parameters": [
          {
            "name": "notificationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Notificationid"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActorResponse": {
        "properties": {
          "userId": {
            "type": "integer",
            "title": "Userid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "profileImageUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profileimageurl"
          }
        },
        "type": "object",
        "required": [
          "userId",
          "name",
          "profileImageUrl"
        ],
        "title": "ActorResponse"
      },
...
      "BetResponse": {
        "properties": {
          "betId": {
            "type": "integer",
            "title": "Betid"
          },
          "todoId": {
            "type": "integer",
            "title": "Todoid"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "requesterId": {
            "type": "integer",
            "title": "Requesterid"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "proofImageUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proofimageurl"
          },
          "isVerified": {
            "type": "boolean",
            "title": "Isverified"
          }
        },
        "type": "object",
        "required": [
          "betId",
          "todoId",
          "content",
          "requesterId",
          "status",
          "proofImageUrl",
          "isVerified"
        ],
        "title": "BetResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NotificationResponse": {
        "properties": {
          "notificationId": {
            "type": "integer",
            "title": "Notificationid"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "actor": {
            "$ref": "#/components/schemas/ActorResponse"
          },
          "todo": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TodoPreviewResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "diaryId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diaryid"
          },
          "bet": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BetResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          },
          "readAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Readat"
          }
        },
        "type": "object",
        "required": [
          "notificationId",
          "type",
          "actor",
          "todo",
          "diaryId",
          "bet",
          "createdAt",
          "readAt"
        ],
        "title": "NotificationResponse"
      },
      "NotificationSettingsRequest": {
        "properties": {
          "discordAlertEnabled": {
            "type": "boolean",
            "title": "Discordalertenabled"
          }
        },
        "type": "object",
        "required": [
          "discordAlertEnabled"
        ],
        "title": "NotificationSettingsRequest"
      },
      "NotificationsPageResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NotificationResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "type": "object",
        "required": [
          "items",
          "nextCursor"
        ],
        "title": "NotificationsPageResponse"
      },
...
      "TodoPreviewResponse": {
        "properties": {
          "todoId": {
            "type": "integer",
            "title": "Todoid"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "todoId",
          "title",
          "description"
        ],
        "title": "TodoPreviewResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}